Added debug camera controls and fixed jump.
Debug camera controls can be used with F1 and WASD to move around the level without using the player character. isKeyJustPressed turns out to be the solution to the problem where consecutive input triggered the increased jump height (for the default jump included in the tutorial code), but this is not a problem for the custom jump. Still, now that the jump uses isKeyJustPressed, it is possible to allow the player to doublejump during the jumping state as well instead of only during the falling state.