Trouble with wall jumping.
Attempted to use rectangles and .overlaps() but while this worked, it mostly just allowed the player to pass through the wall from outside and go into the platform, then once inside the platform, the player would stick to the walls. Decided this was not a good solution if the player would still be able to downjump and upjump through platforms, so attempted to write another method. This method uses the position vector and compares it to the previous frame's position vector, but the problem now is that both of these have the same x position! Testing revealed that when the player is smacked by an enemy and gains x-velocity, the values differ. Apparently, I should implement x-velocity to solve this issue.