Fixed wall jumping.
Player is now able to use the sides of platforms as launching points for more jumps. Player can also jump through a platform to the other side. There is a short time limit on clinging to the side of a platform unless the left or right keys are held down. The player cannot cling to a platform from inside of it.