Smoothed skidding animation and prevented player from skidding off platforms.
There is still an error where the player can get stuck on the side of a platform if it's the right height and the player runs into it from the side. This is because at the point in the code where the player is checking platform collisions for wall climbing, there is a bug where the player is falling due to gravity. Later on in the update code, the player is properly assigned the grounded state. Decided to call it 'skid' instead of 'slide' because later on there may be a separate sliding state for diagonal surfaces.