Separated jump state and iframe state.
This will allow for separate flinch animations when player is grounded vs in the air/being knocked around. A lock state was also added to prevent inputs while being hit (and will be useful for animation locking during attacks). The animation lock period is separate from the iframe period. Animations surrounding flinching are working smoothly now (the flash on damage taken animation is also nicer looking). Damage taken while hanging from a wall will knock the player off the wall. There is one issue remaining: the dust cloud that is supposed to come up when skidding does not appear only when the player is hit and then skids.