Changed attack1 wind down animation, updated position.
Changed the attack1 wind down animation to keep the player in place after attacking. This results in a more natural attack pattern. Also updated the player's position during the attack: the player takes a step forward during the animation and this is now reflected in the player's actual position and hitbox. A debug ShapeRenderer should be used in the future to check attack and player hitboxes during animations. There is still unintended behavior in when the player is freed from the attack animation lock: the player should be allowed to give inputs again after the downswing of attack1 is finished, but currently it seems there is some delay before inputs are accepted again. Part of the wind down animation of attack1 is able to play even if mashing other input buttons to interrupt it.