Fixed bugs.
Player no longer gets stuck on platforms that are just the right height and width when running by them and not attempting to stick to them. When player is attacking and the position change moves them off a platform, the player will fall properly instead of waiting until the animation is complete to fall. The player is no longer locked for an unintended amount of time after attacking: running and other actions are available immediately before the attack1 wind down animation begins to play.