Added a boost up to platform.
Player can press up to climb onto a platform in range. This is useful for just barely missing a platform on a jump: the player can course correct while passing through the platform. Also fixed damage knocking the player off walls, but currently the player goes into a damage lock state and immediately after that ends, if the player is still falling they can move midair in a way that looks unnatural.