Fixed bugs, added dodge.
Fixed bug where swapping to the wrong arrow key very quickly within a certain time could allow the player to continue hanging on the wall despite not pressing the key that should cause the player to face the wall. The player was not facing the right direction according to input. Also fixed the issue where after being knocked off a wall by damage, the player could hold the arrow key for the direction opposite of where the player was flying to, and after the input lock from taking damage wore off, the player would unnaturally change directions midair. Also added a dodge with included invincibility period.