Added base item class, fixed bugs.
Fixed bug that allowed player to stick to wall even if pressing the wrong arrow key (This time, for good. Probably.). Added base item class that has code in place for animating and is lootable, but currently does not do anything on loot and for some reason the animations do not currently update.