Moved keys to variables, added hit effect, added squat.
Moved keys to variables in preparation for giving player ability to rebind controls as they see fit. Hit effects are now functional both when the player is damaged by a monster and when the player damages a monster. When the player presses down, an appropriate sprite is shown and the player's hitbox height is smaller. This allows players to duck under projectiles.