Added slash hit effect, dodge iframes, dodge dust.
Since the current attack animation uses the blade of the shovel to hit enemies, a slash hit effect was added. In the future, the impact hit effect will be used if the blunt side of the shovel is used in the attack animation. Player can now ignore sources of damage when dodging through them. Dodge produces a dust cloud similar to skidding. Dodge currently uses the skidding animation.