Added damage for enemies/projectiles, damage effects.
Sources of damage now have damage values associated with them. Damage effects have been polished and numbers fade out and fly around in an arc depending on the direction of the incoming damage. This applies for both the player and enemies. The boosting up to platform animation was changed so the player has a knockback-ignoring state while performing this action.