Removed items from entity list.
Entities will now be interactables that can damage or do damage. Next on the to-do list: add attack damage zones to the debug renderer. Also, the dandelion enemy aims projectiles at the center of the player's hitbox instead of at the player's position. This prevents the player from being able to crouch and ignore all projectiles forever.