Breakable objects are now enemies and can be destroyed.
Removed breakable object class because it was unnecessary. Instead, enemies have a touch damage boolean and some enemies are just objects that cannot do touch damage or move. Eventually, there will be more overlap. Some enemies will not be able to move but will be able to do damage or perform other actions.