Hit effects updated & centered on hitboxes.
Moved hit effect spawn to center on hitboxes within a range instead of centering on position within a range. This should result in better placement. Also added more polished effect sprites and layering of different types of sprites for slash effects. Hit effects now face a random direction to add chaos unless it makes sense for them to be oriented to a specific direction (as in the case of piercing projectile damage).