Enemies now have sight ranges.
Prepwork for enemy AI laid down. The dandelion enemy now only shoots the player if the player's hitbox is overlapping with the enemy's range of sight box. Enemies have range of sight boxes that render in the debug renderer. Progress was also made on the graphics mockup since the lack of a design document is causing stalling on the graphics front.