Minor refactoring. Basic enemy can now return to home platform and jump.
Improved enemy AI to act more natural. Enemy no longer vibrates in place when on top of the player. After chasing the player away from its home platform, if the enemy loses sight of the player, it will navigate back to its home platform and do a 'superjump' or a downjump to return there. TODO: wander/idle ai, give the enemy an 'overshooting' behavior so it isn't so robotic when attacking the player. Currently, it moves to the player until it is touching them and then it stops moving completely until the player is pushed away by touch damage.