Improved enemy behavior.
Enemy now performs various randomized wander actions when not chasing a player, and when chasing a player, the enemy can jump and downjump properly to reach the target. Basic enemy floating beetle is now displayed with a little bob animation. Enemy speed is adjusted when targeting a player vs when wandering. Still need to give enemy 'overshooting' behavior.