Health orb is now interactable.
Click the health orb to swap to the setting that keeps the player health displayed on the orb. Click again to revert back to not displaying the health number. If the setting is not on, then the player can hover over the orb to temporarily view the health number. The number will vanish when the mouse is no longer hovering over the orb. Planning to use a water shader for health liquid inside orb.