Added life item, life gaining, fixed bugs.
The large heal item now has a different placeholder image. There is a new life item that uses the old heart placeholder image. When this life item is looted, it gives the player a life if the player does not already have the maximum possible lives. There is a little growing animation for the petal when the player gains a life and the loss of life animation has been tweaked a bit so the petal looks 2D. Fixed a bug that was causing the beetle enemy to respawn in the wrong position and then snap to the correct position.