Edited level bounds, enemies stay in bounds, no respawn on permadeath.
Level width boundary is now by default determined by the width of the lowest floor platform on the map. Still need to prevent camera from moving out of bounds of the map and displaying empty space. Death is fixed so the player does not respawn when out of lives. The flashing effect does not appear if the player will not respawn. The life petal indicator was crashing the game when the player ran out of lives and died again, but that is now fixed. Enemies also now stay in level boundaries.