Camera stays in level bounds now.
Prevent player from seeing blank empty space outside of level bounds. The camera follows player as normal but hits a half screen from level bounds limit and does not move further. The camera can move slightly further on the y-axis for the bottom of the map so there will be room for foreground graphics.