Improved Wasp Scout behavior.
Scout now flies close enough to player to be visible on default window size, but still remains out of attack range. Horizontal flight speed was changed to improve flight smoothness when following player. Vertical flight speed was increased to prevent the player from falling out of aggroRange so easily. Scout also puts on a burst of speed if the player attempts to dodge out of the aggroRange box. This allows the Wasp Scout to stay in range more easily. Parameters were also changed so the Scout keeps more of the aggroRange box overlapping with the player hitbox at any time. This allows the Scout to more easily keep the player in sight.