Edited enemy chase behavior.
When enemy does not need to move in aggro mode (basically, enemy is standing on the player's hitbox or x coordinate), it now moves around left and right a random amount instead of standing still, vibrating in place, or moving back and forth over the player's hitbox. Still need to figure out how to get enemy to behave this way when player is pressed against the level boundary.