Wasp scout can now call other wasps.
Added abstract wasp class and EnemyType enum. EnemyType will be used in the future to regulate interactions between different groups of enemies. Wasp scout can now call other wasps on the map to the player's current location and will attempt to do this immediately on seeing player exactly once. May be a good idea to change the counter to reset if the scout loses sight of the player and exits aggro mode. Will create other wasps to test the call method.