Enemies go home after player dies.
Also, there is a bug where the player vibrates a little when being pushed against the left wall of the map but not the right wall. Currently, there is some code that prevents the player from vibrating when dying against the left wall, but the player still vibrates when being hit against the left wall. The code for dying against the wall will be unnecessary when the bug is fixed.