Fixed knockdown state, refactored damage taking for player.
Player is now knocked down properly with animation. Player is not given iframes when knocked down, only when damaged. Damage refactored so enemies detect when to attack player. Currently, there is a bug where any animation that is interrupted (happens often when being knocked down) will not reset the player's HitState. This causes various problems.