Fixed bugs.
Dodge dust cloud was only spawning every other dodge. Now it spawns on every dodge as intended. This was because the jump counter was not resetting until the player was out of jumps (and this had the side effect of sometimes causing the player to be unable to doublejump even after launching off a surface). Also fixed death and down state sliding dust cloud. Fixed crash caused by level looping through renderables render methods and player render method calling spawnDustCloud in the level. This was a ConcurrentModificationException. Fixed crash caused by entities being added to the renderables list before their eyeHeight was assigned. Entities rely on the eyeHeight value to provide the renderables sorting method with a y comparison value. Fixed render order so that player renders under grass on their platform but over grass on higher platforms that they overlap (same with other entities, such as wasps). There is still a bug where sometimes a wasp gets stuck at the return to home platform point.