Finished adding renderables to renderables list.
Still using a list for different types of objects, will phase that out. Also need to add an updatables interface. When the old list is phased out, need to go back and replace all the level arguments so that level creates things and adds them to renderables list itself. Also found a bug where hitting the wasp scout seems to confuse it, causing it to stop moving.